typedarray-memory refactor for naive memory implementation done
3 memory types
  - naive: it has a list of pointers which point to an index in
           the data array. if the pointer value is zero, it has
           been freed.
           when we run out of space, a new data array is created,
           and all non freed values are copied over and their
           pointers updated
  - const: same as naive, can't be freed
  - stack: same as naive, but free frees up everything starting
           from the given pointer, not just that one value to
           which the pointer is pointing to

additional note: neither of the memory types actually know the
size of any of the segments allocated. all they know is where
each of them start and what's the next free data index.
this is because sometimes I want to resize a number to have
smaller length than I've allocated and I want to be able
to do this without allocating a new one. If I store the size
of the numbers in the actual allocated data, it's one
simple write to shrink the number's length, instead of a whole
new allocate.

also, when resizing the data, values are copied by the stored size
which means only the actual size will be used.

of course this means there is a hidden assumption in the resize
algorithm, that the **first** value is the size for the segment.

I intend to change this to work with a user given solution, or at
least make it more obvious. Perhaps store the length when allocate
and then allow to be changed.